materialize the `{APP_DATA_DIR}/storage` location for Prisma's use
Placing a file in the app's destination volume for the SQLite database
the app needs to create appears to force Tipi to materialize the folder
at a point in the process where AnythingLLM's Prisma client could access
it on first run.

Without this, the Prisma client tried to create the file at a
as-yet-nonexistant path and failed. Tipi subsequently created the volume
location (b/c it was there on the host after Prisma failed) and a simple
app restart allowed Prisma to run its migrations on the file it could
now create.

steps:

- attempt an init service
- try getting rid of the env var for the storage and stop depending on
the init service
- try an init service again; removing the env var did nothing
- materialize the file?
- remove the volume init service
- try nesting the storage volume inside data
- along with precreating the file? or maybe i won't need that?
- try without preseeding the file
- what if we materialize a different file?
- non-debugging `restart` flag