fix: remove quotes and add credentials to Planka
The quotes were included the environment variable which caused the
application to fail to boot up with a "getaddrinfo EAI_AGAIN base"
error.
Planka seems to not have a default admin unless it's specified so I
copied the lines from
https://github.com/plankanban/planka/blob/master/docker-compose.yml.
After these changes, I was able to boot Planka from Tipi.